hwdef: skyviper-journey delete non-existent params
The profile parameters seem to have been part of SkyRocketToys' fork and
never made it into upstream. I think the same is true for the FLIP
parameters and some of the other TMODE parameters.

SYSID_SW_TYPE does nothing anymore, and I'm not sure why it was ever
set to 100 in the first place. Must have been some fork-specific thing
that we don't care about.